+++ src/+aod/+persistent/+events/HdfPathEvent.m
@@ -1,0 +1,16 @@
+classdef (ConstructOnLoad) HdfPathEvent < event.EventData 
+
+    properties 
+        Entity
+        NewPath 
+        OldPath 
+    end
+
+    methods 
+        function obj = HdfPathEvent(entity, oldPath, newPath)
+            obj.Entity = entity;
+            obj.OldPath = oldPath;
+            obj.NewPath = newPath;
+        end
+    end
+end