+++ src/+aod/+persistent/EntityFactory.m
@@ -1,5 +1,5 @@
 classdef EntityFactory < handle 
-% ENTITYFACTORY
+% Middle layer between persistent interface and HDF5 file
 %
 % Description:
 %   A factory for creating entities from an HDF5 file and caching the 
@@ -11,14 +11,17 @@ classdef EntityFactory < handle
 % Static method access:
 %   experiment = EntityFactory.init(hdfName)
 
-% By Sara Patterson, 2022 (AOData)
+% By Sara Patterson, 2023 (AOData)
 % -------------------------------------------------------------------------
 
     properties (SetAccess = private)
-        hdfName
-        persistor
-        entityManager
-        classManager
+        hdfName                 string 
+        % Handles changes made from persistent interface to HDF5 file
+        persistor               aod.persistent.Persistor
+        % Tracks all entities in the HDF5 file
+        entityManager           aod.h5.EntityManager 
+        classManager            aod.infra.ClassRepository
+        listeners               event.listener
         cache 
     end
 
@@ -28,13 +31,34 @@ classdef EntityFactory < handle
             obj.entityManager = aod.h5.EntityManager(hdfName);
             obj.classManager = aod.infra.ClassRepository();
             obj.persistor = aod.persistent.Persistor(hdfName);
-            addlistener(obj.persistor, 'EntityChanged', @obj.onEntityChanged);
+            
+            obj.listeners = [...
+                addlistener(obj.persistor, 'EntityChanged', @obj.onEntityChanged);...
+                addlistener(obj.persistor, 'HdfPathChanged', @obj.onEntityRenamed)];
 
             obj.cache = containers.Map();
         end
 
         function e = getExperiment(obj)
             e  = obj.create('/Experiment');
+        end
+
+        function uuid = path2uuid(obj, hdfPath)
+            T = obj.entityManager.table;
+            idx = find(T.Path == hdfPath);
+            if isempty(idx)
+                uuid = [];
+            else
+                uuid = T.UUID(idx);
+            end
+        end
+
+        function p = isCached(obj, UUID)
+            if obj.cache.isKey(UUID)
+                p = obj.cache(UUID);
+            else
+                p = [];
+            end
         end
 
         function clearCache(obj)
@@ -56,6 +80,7 @@ classdef EntityFactory < handle
             className = T{T.Path == hdfPath, 'Class'};
             entityType = aod.core.EntityTypes.get(obj.entityManager.entityMap(uuid));
 
+            %! Consider deprecating
             mirrorFcn = str2func(aod.infra.findMirror(...
                 entityType, className, obj.classManager));
             e = mirrorFcn(obj.hdfName, hdfPath, obj);
@@ -67,7 +92,7 @@ classdef EntityFactory < handle
 
     methods (Access = private)
         function onEntityChanged(obj, ~, evt)
-            % Callback for entity changes
+            % Update the middle layer to reflect changes to an entity
             % -------------------------------------------------------------
 
             % Refresh the list of UUIDs
@@ -76,6 +101,77 @@ classdef EntityFactory < handle
             % Make sure irrelevant UUIDs are removed from the cache
             if strcmp(evt.Action, 'Remove')
                 remove(obj.cache, evt.UUID);
+            end
+        end
+
+        function onEntityRenamed(obj, ~, evt)
+            % Handles side-effects of a renamed group and the path change
+            %
+            % HDF5 group renaming is a nightmare and here I've chosen to 
+            % minimize disruption to user over minimizing complexity
+            % -------------------------------------------------------------
+
+            % Refresh the list of UUIDs
+            obj.entityManager.collect();
+            T = obj.entityManager.table;
+
+            % Update the entity's HDF5 path
+            evt.Entity.changeHdfPath(evt.NewPath);
+            evt.Entity.reload();
+
+            % Get the modified entity and all children
+            idx = startsWith(T.Path, evt.NewPath);
+            allPaths = T{idx, "Path"};
+            allUUIDs = T{idx, "UUID"};
+
+            % Log all HDF paths with an updated HDF5 path, for reload
+            updatedEntities = string.empty();
+            % Change the HDF path of all cached entities (original + child)
+            % Entities that are not cached will be loaded with correct path
+            for i = 1:numel(allPaths)
+                entity = obj.isCached(allUUIDs(i));
+                if ~isempty(entity)
+                    entity.changeHdfPath(strrep(entity.hdfPath, evt.OldPath, evt.NewPath));
+                    updatedEntities = cat(1, updatedEntities, entity.hdfPath);
+                end
+            end
+
+            % Update entity's parent containers to reflect new HDF5 paths
+            evt.Entity.Parent.reload();
+
+            % Collect all softlinks. Softlink sources are changed by 
+            % H5L.move, but softlink targets are not and must be edited
+            links = aod.h5.collectExperimentLinks(obj.hdfName, true);
+            for i = 1:numel(allPaths)
+                % Check if original HDF path is a softlink target
+                idx = find(links.Target == strrep(allPaths(i), evt.NewPath, evt.OldPath));
+                if isempty(idx)
+                    continue 
+                end
+                sourcePaths = links{idx, "Location"};
+                % Entities cannot be loaded until link is fixed so correct
+                % through h5tools rather than persistent interface 
+                for j = 1:numel(sourcePaths)
+                    %linkAttr = h5tools.readatt(obj.hdfName, sourcePaths(j), 'all');
+                    h5tools.deleteObject(obj.hdfName, sourcePaths(j));
+                    [sourcePath, sourceLink] = h5tools.util.splitPath(sourcePaths(j));
+                    h5tools.writelink(obj.hdfName, sourcePath, sourceLink, allPaths(i));
+                    %h5tools.writeatt(obj.hdfName, sourcePaths(j), linkAttr);
+                    updatedEntities = cat(1, updatedEntities, sourcePaths);
+                end
+            end
+
+            % Now we need to reload any cached entities with a changed 
+            % softlink - reload is preferable to deletion because user may 
+            % have existing handles to these entities
+            updatedEntities = unique(updatedEntities);
+            for i = 1:numel(updatedEntities)
+                uuid = obj.path2uuid(updatedEntities(i));
+                e = obj.isCached(uuid);
+                if isempty(e)
+                    continue 
+                end
+                e.reload();
             end
         end
     end